Implement player positioning on serve

+++ src/main/java/com/ft/emulator/server/game/core/constants/GameFieldSide.java
@@ -1,6 +1,6 @@
 package com.ft.emulator.server.game.core.constants;
 
 public class GameFieldSide {
-    public final static int RedTeam = 0;
-    public final static int BlueTeam = 1;
+    public final static short RedTeam = 0;
+    public final static short BlueTeam = 1;
 }
